requirements.txt
asttokens==3.0.0
certifi==2022.12.7
charset-normalizer==2.1.1
colorama==0.4.6
comm==0.2.2
debugpy==1.8.14
decorator==5.2.1
exceptiongroup==1.3.0
executing==2.2.0
filelock==3.13.1
idna==3.4
ipykernel==6.29.5
ipython==8.37.0
jedi==0.19.2
Jinja2==3.1.3
jupyter_client==8.6.3
jupyter_core==5.8.1
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.3
numpy==2.1.2
packaging==25.0
parso==0.8.4
pillow==11.0.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
psutil==7.0.0
pure_eval==0.2.3
Pygments==2.19.1
python-dateutil==2.9.0.post0
pywin32==310
pyzmq==26.4.0
requests==2.28.1
six==1.17.0
stack-data==0.6.3
sympy==1.13.3
torch==2.0.1+cu117
torchvision==0.15.2+cu117
tornado==6.5.1
traitlets==5.14.3
typing_extensions==4.9.0
urllib3==1.26.13
wcwidth==0.2.13
